% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-a-fortran.R
\name{call_a_fortran}
\alias{call_a_fortran}
\title{R wrapper function calling a Fortran subroutine}
\usage{
call_a_fortran(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
double the elements of the numeric vector
}
\description{
R wrapper function calling a Fortran subroutine
}
